refactor: ShadCN CLI path
