+++ app/components/ui/tiptap/TiptapIcon.vue
