+++ app/components/ui/tiptap/TiptapMobileToolbar.vue
