+++ app/components/ui/tiptap/TiptapSlotPanel.vue
