+++ app/components/ui/tiptap/TiptapStatusBar.vue
