+++ app/components/ui/tiptap/TiptapTreeItem.vue
